basic popup working

+++ inc/schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS `uri_data` (
   `log` text NOT NULL,
   `modified` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
   `created` timestamp NOT NULL DEFAULT '0000-00-00 00:00:00',
-  `stale` tinyint(4) NOT NULL,
+  `stale` int(11) NOT NULL,
   PRIMARY KEY (`id`),
   UNIQUE KEY `uri` (`uri`),
   FULLTEXT (words)
